requirements.txt
# Django Core
Django>=5.2.3,<5.3

# Database - PostgreSQL
psycopg2-binary>=2.9.0

# Image Processing
Pillow>=10.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Static Files
whitenoise>=6.0.0

# Environment Variables
python-dotenv>=1.0.0

# WSGI Server for Production
gunicorn>=21.0.0

# Security
cryptography>=41.0.0
django-csp==3.8

# Database URL parsing
dj-database-url>=2.0.0

# Django REST Framework
djangorestframework>=3.14.0

# Django CORS Headers
django-cors-headers>=4.0.0

# Django Filter
django-filter>=23.0

# Django Tailwind
django-tailwind>=3.4.0

# Additional security packages
django-environ>=0.10.0

# Payment Gateway
requests>=2.31.0
zarinpal>=1.0.0
zarinpal-py-sdk>=0.1.4  # SDK رسمی زرین‌پال

# SMS Provider
kavenegar
ippanel>=2.0.7  # فراز اس‌ام‌اس SDK رسمی

# Jalali Date
jdatetime>=4.1.0
